# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing a city simulation game",
        "title": "City Management API",
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/city": {
            "get": {
                "description": "Get current city state and information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "city"
                ],
                "summary": "Get city information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.City"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new city with specified name and difficulty",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "city"
                ],
                "summary": "Create a new city",
                "parameters": [
                    {
                        "description": "City creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateCityRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.City"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/city/citizens": {
            "get": {
                "description": "Get list of all citizens in the city",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "citizens"
                ],
                "summary": "List all citizens",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Citizen"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new citizen with specified name and job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "citizens"
                ],
                "summary": "Create a new citizen",
                "parameters": [
                    {
                        "description": "Citizen creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CitizenCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Citizen"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/city/citizens/jobs": {
            "get": {
                "description": "Get statistics of citizens by job type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "citizens"
                ],
                "summary": "Get job statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.JobsStatsResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/city/citizens/mass-add": {
            "post": {
                "description": "Add multiple citizens with specified job and prefix",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "citizens"
                ],
                "summary": "Mass add citizens",
                "parameters": [
                    {
                        "description": "Mass add request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MassAddRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Citizen"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/city/citizens/{id}": {
            "delete": {
                "description": "Delete a citizen by ID",
                "tags": [
                    "citizens"
                ],
                "summary": "Delete a citizen",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Citizen ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/city/citizens/{id}/happiness": {
            "patch": {
                "description": "Adjust a citizen's happiness by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "citizens"
                ],
                "summary": "Change citizen happiness",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Citizen ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Happiness change request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ChangeHappinessRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Citizen"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/city/citizens/{id}/job": {
            "patch": {
                "description": "Change a citizen's job by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "citizens"
                ],
                "summary": "Change citizen job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Citizen ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Job change request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ChangeJobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Citizen"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/city/settings": {
            "patch": {
                "description": "Update city configuration settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "city"
                ],
                "summary": "Update city settings",
                "parameters": [
                    {
                        "description": "Settings update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CitySettings"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Building": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "health": {
                    "type": "integer",
                    "example": 100
                },
                "id": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "models.ChangeHappinessRequest": {
            "type": "object",
            "properties": {
                "delta": {
                    "type": "number",
                    "example": 5,
                    "minimum": -100,
                    "maximum": 100
                }
            }
        },
        "models.ChangeJobRequest": {
            "type": "object",
            "properties": {
                "job": {
                    "type": "string",
                    "example": "worker",
                    "enum": [
                        "unemployed",
                        "farmer",
                        "worker",
                        "engineer"
                    ]
                }
            }
        },
        "models.Citizen": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "happiness": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "job": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.CitizenCreateRequest": {
            "type": "object",
            "properties": {
                "job": {
                    "type": "string",
                    "example": "farmer",
                    "enum": [
                        "unemployed",
                        "farmer",
                        "worker",
                        "engineer"
                    ]
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                }
            }
        },
        "models.City": {
            "type": "object",
            "properties": {
                "buildings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Building"
                    }
                },
                "citizens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Citizen"
                    }
                },
                "day": {
                    "type": "integer"
                },
                "resources": {
                    "$ref": "#/definitions/models.Resources"
                },
                "settings": {
                    "$ref": "#/definitions/models.CitySettings"
                }
            }
        },
        "models.CitySettings": {
            "type": "object",
            "properties": {
                "baseFoodUsePerCap": {
                    "type": "integer"
                },
                "difficulty": {
                    "type": "string"
                },
                "energyPerEngineer": {
                    "type": "integer"
                },
                "foodPerFarmer": {
                    "type": "integer"
                },
                "happinessDecay": {
                    "type": "number"
                },
                "moneyPerWorker": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.CreateCityRequest": {
            "type": "object",
            "properties": {
                "difficulty": {
                    "type": "string",
                    "example": "normal",
                    "enum": [
                        "easy",
                        "normal",
                        "hard"
                    ]
                },
                "name": {
                    "type": "string",
                    "example": "My City"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "city not found"
                }
            }
        },
...
        "models.JobsStatsResponse": {
            "type": "object",
            "properties": {
                "engineer": {
                    "type": "integer",
                    "example": 1
                },
                "farmer": {
                    "type": "integer",
                    "example": 5
                },
                "unemployed": {
                    "type": "integer",
                    "example": 2
                },
                "worker": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "models.MassAddRequest": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 5,
                    "minimum": 1,
                    "maximum": 100
                },
                "job": {
                    "type": "string",
                    "example": "farmer",
                    "enum": [
                        "unemployed",
                        "farmer",
                        "worker",
                        "engineer"
                    ]
                },
                "prefix": {
                    "type": "string",
                    "example": "Citizen"
                }
            }
        },
...
        "models.Resources": {
            "type": "object",
            "properties": {
                "energy": {
                    "type": "integer"
                },
                "food": {
                    "type": "integer"
                },
                "money": {
                    "type": "integer"
                }
            }
        },
...
        "models.UpdateSettingsRequest": {
            "type": "object",
            "properties": {
                "baseFoodUsePerCap": {
                    "type": "integer",
                    "example": 2
                },
                "difficulty": {
                    "type": "string",
                    "example": "hard",
                    "enum": [
                        "easy",
                        "normal",
                        "hard"
                    ]
                },
                "energyPerEngineer": {
                    "type": "integer",
                    "example": 8
                },
                "foodPerFarmer": {
                    "type": "integer",
                    "example": 12
                },
                "happinessDecay": {
                    "type": "number",
                    "example": 0.7
                },
                "moneyPerWorker": {
                    "type": "integer",
                    "example": 10
                },
                "name": {
                    "type": "string",
                    "example": "New City Name"
                }
            }
        }
    }
}